requirements.txt
elevenlabs>=1.0.0
requests
speechrecognition
sounddevice
soundfile
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
flask>=2.3.0
flask-cors>=4.0.0
waitress>=2.1.2